feat: extract gateway_notes package (service + schemas)
Move NotesService and the notes response schemas (NoteResponse,
CreateNoteRequest, CreateNoteResponse, NotesPageResponse) out of
gateway_oss into a new gateway_notes package that depends only on
gateway_core. gateway_oss re-exports both through thin facades
(services/notes.py, models/schemas.py) so existing REST/MCP call
sites keep working unchanged. NotesService also gains a
get_notes_for_profile method mirroring PostsService's (temporary,
intentional duplication until profiles migrates to gateway_notes).
