+++ packages/gateway_notes/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "gateway_notes"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "pydantic>=2.13.4"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_notes"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
