+++ packages/gateway_oss/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
   "tenacity>=9.1.2",
   "uvloop>=0.22.1",
   "gateway_core",
+  "gateway_notes",
 ]
 
 [build-system]
@@ -25,3 +26,4 @@ packages = ["src/gateway_oss"]
 
 [tool.uv.sources]
 gateway_core = { workspace = true }
+gateway_notes = { workspace = true }
