+++ packages/gateway_oss/src/gateway_oss/models/schemas.py
@@ -1,10 +1,14 @@
 from __future__ import annotations
 
-import logging
-
 from pydantic import BaseModel
 
 from gateway_core.auth import BearerCredentials  # noqa: F401
+from gateway_notes.schemas import (  # noqa: F401
+    CreateNoteRequest,
+    CreateNoteResponse,
+    NoteResponse,
+    NotesPageResponse,
+)
 from gateway_oss.converters.markdown import (
     html_to_markdown,
 )
@@ -12,15 +16,10 @@ from gateway_oss.models.substack import (
     SubstackComment,
     SubstackFollowingUser,
     SubstackFullPost,
-    SubstackNote,
-    SubstackNoteCreated,
-    SubstackNotesPage,
     SubstackPreviewPost,
     SubstackProfilePostsPage,
     SubstackPublicProfile,
 )
-
-_log = logging.getLogger(__name__)
 
 
 class LivenessResponse(BaseModel):
@@ -52,69 +51,6 @@ class ProfileResponse(BaseModel):
             url=f"https://substack.com/@{profile.handle}",
             avatar_url=profile.photo_url or "",
             bio=profile.bio,
-        )
-
-
-# ------------------------------------------------------------------
-# Notes
-# ------------------------------------------------------------------
-
-
-class NoteAuthor(BaseModel):
-    id: int
-    name: str
-    handle: str
-    avatar_url: str
-
-
-class NoteResponse(BaseModel):
-    id: int
-    body: str
-    likes_count: int
-    author: NoteAuthor | None = None
-    published_at: str
-
-    @classmethod
-    def from_substack(cls, note: SubstackNote) -> NoteResponse:
-        user = note.context.users[0] if note.context.users else None
-        comment = note.comment
-        if comment is None:
-            _log.warning(
-                "Note %r has no comment body; returning empty defaults", note.entity_key
-            )
-        if user is None:
-            _log.warning(
-                "Note %r has no author; returning empty defaults", note.entity_key
-            )
-        return cls(
-            id=comment.id if comment else 0,
-            body=comment.body if comment else "",
-            likes_count=comment.reaction_count
-            if (comment and comment.reaction_count is not None)
-            else 0,
-            author=(
-                NoteAuthor(
-                    id=user.id,
-                    name=user.name,
-                    handle=user.handle,
-                    avatar_url=user.photo_url or "",
-                )
-                if user
-                else None
-            ),
-            published_at=note.context.timestamp,
-        )
-
-
-class NotesPageResponse(BaseModel):
-    items: list[NoteResponse]
-    next: str | None = None
-
-    @classmethod
-    def from_substack(cls, page: SubstackNotesPage) -> NotesPageResponse:
-        return cls(
-            items=[NoteResponse.from_substack(n) for n in page.items],
-            next=page.next_cursor,
         )
 
 
@@ -241,21 +177,3 @@ class FollowingResponse(BaseModel):
     @classmethod
     def from_substack(cls, users: list[SubstackFollowingUser]) -> FollowingResponse:
         return cls(items=[FollowingUserResponse.from_substack(u) for u in users])
-
-
-# ------------------------------------------------------------------
-# Note creation
-# ------------------------------------------------------------------
-
-
-class CreateNoteRequest(BaseModel):
-    content: str
-    attachment: str | None = None
-
-
-class CreateNoteResponse(BaseModel):
-    id: int
-
-    @classmethod
-    def from_substack(cls, note: SubstackNoteCreated) -> CreateNoteResponse:
-        return cls(id=note.id)
